requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic-settings==2.2.1
google-genai==0.3.0
twilio==8.12.0
python-dotenv==1.0.1
apscheduler==3.10.4
